Refactor the existing resource configuration into separate file and make code clearer
